# Fix Zlib performance regression