--- a/lineproto/reader_test.go
+++ b/lineproto/reader_test.go
@@ -21,7 +21,8 @@ func TestReader(t *testing.T) {
 	require.NoError(t, err)
 	require.Equal(t, l1, l1Expected)
 
-	r.ActivateZlib()
+	err = r.ActivateZlib()
+	require.NoError(t, err)
 
 	l2Expected := []byte("$OtherCommand test|")
 	l2, err := r.ReadLine()