Fix Zlib performance regression by manually buffering data. Fixes #9 

Performance comparison with the latest version:
```
benchmark             old ns/op     new ns/op     delta
BenchmarkReader-4     93300098      11941476      -87.20%

benchmark             old allocs     new allocs     delta
BenchmarkReader-4     7              6              -14.29%

benchmark             old bytes     new bytes     delta
BenchmarkReader-4     35372         4374          -87.63%
```

Comparison with the previous version (pre #8):
```
benchmark             old ns/op     new ns/op     delta
BenchmarkReader-4     13904844      11941476      -14.12%

benchmark             old allocs     new allocs     delta
BenchmarkReader-4     6              6              +0.00%

benchmark             old bytes     new bytes     delta
BenchmarkReader-4     4950          4374          -11.64%
```

<!-- Reviewable:start -->
---
This change is [<img src="https://reviewable.io/review_button.svg" height="34" align="absmiddle" alt="Reviewable"/>](https://reviewable.io/reviews/direct-connect/go-dc/10)
<!-- Reviewable:end -->